requirements.txt
streamlit
openai
python-dotenv
playwright
langchain
speechrecognition
pyaudio
